requirements.txt
aniso8601==7.0.0
astroid==2.5.1
colorama==0.4.4
Django==2.2.4
django-cors-headers==3.7.0
django-filter==2.4.0
django-graphql-jwt==0.3.1
graphene==2.1.8
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
isort==5.7.0
lazy-object-proxy==1.5.2
mccabe==0.6.1
promise==2.3
PyJWT==2.0.1
pylint==2.7.2
pytz==2021.1
qrcode==6.1
Rx==1.6.1
singledispatch==3.6.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
wrapt==1.12.1